fix: Optimize backup logic for cronjob

+++ agent/app/service/cronjob_helper.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/1Panel-dev/1Panel/agent/utils/alert_push"
+	"github.com/pkg/errors"
 
 	"github.com/1Panel-dev/1Panel/agent/app/dto"
 	"github.com/1Panel-dev/1Panel/agent/app/model"
@@ -191,14 +192,14 @@ func (u *CronjobService) handleNtpSync(cronjob model.Cronjob, taskItem *task.Tas
 }
 
 func (u *CronjobService) handleCutWebsiteLog(cronjob *model.Cronjob, startTime time.Time, taskItem *task.Task) error {
+	clientMap := NewBackupClientMap([]string{fmt.Sprintf("%v", cronjob.DownloadAccountID)})
+	if !clientMap[fmt.Sprintf("%d", cronjob.DownloadAccountID)].isOk {
+		return errors.New(i18n.GetMsgWithDetail("LoadBackupFailed", clientMap[fmt.Sprintf("%d", cronjob.DownloadAccountID)].message))
+	}
 	taskItem.AddSubTaskWithOps(i18n.GetWithName("CutWebsiteLog", cronjob.Name), func(t *task.Task) error {
 		websites := loadWebsForJob(*cronjob)
 		fileOp := files.NewFileOp()
 		baseDir := GetOpenrestyDir(SitesRootDir)
-		clientMap, err := NewBackupClientMap([]string{fmt.Sprintf("%v", cronjob.DownloadAccountID)})
-		if err != nil {
-			return fmt.Errorf("load local backup client failed, err: %v", err)
-		}
 		for _, website := range websites {
 			taskItem.Log(website.Alias)
 			var record model.BackupRecord
@@ -264,34 +265,6 @@ func (u *CronjobService) handleSystemClean(cronjob model.Cronjob, taskItem *task
 	taskItem.AddSubTaskWithOps(i18n.GetMsgByKey("HandleSystemClean"), cleanTask, nil, int(cronjob.RetryTimes), time.Duration(cronjob.Timeout)*time.Second)
 }
 
-func (u *CronjobService) uploadCronjobBackFile(cronjob model.Cronjob, task *task.Task, accountMap map[string]backupClientHelper, file string) (string, error) {
-	defer func() {
-		_ = os.Remove(file)
-	}()
-	var errItem error
-	accounts := strings.Split(cronjob.SourceAccountIDs, ",")
-	cloudSrc := strings.TrimPrefix(file, global.Dir.LocalBackupDir+"/tmp/")
-	for _, account := range accounts {
-		if len(account) != 0 {
-			task.LogStart(i18n.GetMsgWithMap("UploadFile", map[string]interface{}{
-				"file":   pathUtils.Join(accountMap[account].backupPath, cloudSrc),
-				"backup": accountMap[account].name,
-			}))
-			_, err := accountMap[account].client.Upload(file, pathUtils.Join(accountMap[account].backupPath, cloudSrc))
-			task.LogWithStatus(
-				i18n.GetMsgWithMap("UploadFile", map[string]interface{}{
-					"file":   pathUtils.Join(accountMap[account].backupPath, cloudSrc),
-					"backup": accountMap[account].name,
-				}), err)
-			if err != nil {
-				errItem = err
-				continue
-			}
-		}
-	}
-	return cloudSrc, errItem
-}
-
 func (u *CronjobService) removeExpiredBackup(cronjob model.Cronjob, accountMap map[string]backupClientHelper, record model.BackupRecord) {
 	var opts []repo.DBOption
 	opts = append(opts, repo.WithByFrom("cronjob"))
@@ -314,6 +287,9 @@ func (u *CronjobService) removeExpiredBackup(cronjob model.Cronjob, accountMap m
 					if _, ok := accountMap[account]; !ok {
 						continue
 					}
+					if !accountMap[account].isOk {
+						continue
+					}
 					_, _ = accountMap[account].client.Delete(pathUtils.Join(accountMap[account].backupPath, "system_snapshot", records[i].FileName))
 				}
 			}
@@ -322,6 +298,9 @@ func (u *CronjobService) removeExpiredBackup(cronjob model.Cronjob, accountMap m
 			for _, account := range accounts {
 				if len(account) != 0 {
 					if _, ok := accountMap[account]; !ok {
+						continue
+					}
+					if !accountMap[account].isOk {
 						continue
 					}
 					_, _ = accountMap[account].client.Delete(pathUtils.Join(accountMap[account].backupPath, records[i].FileDir, records[i].FileName))
